Fix mapping errors
